01-smart-image-analyzer/requirements.txt
torch
torchvision
opencv-python
matplotlib
pillow
ultralytics